+++ api/Dockerfile
@@ -2,13 +2,12 @@ FROM python:3.9
 
 WORKDIR /api
 
+COPY src src
 COPY requirements.txt .
-COPY app.py .
-COPY algorithms.py .
 COPY .env .
 
 RUN pip3 install -r requirements.txt
 
-CMD ["python3", "app.py"]
+CMD ["python3", "src/main.py"]
 
 EXPOSE 3020
