+++ api/src/main.py
@@ -1,0 +1,14 @@
+from flask import Flask
+from api import api
+from settings import load_config
+from flask_cors import CORS
+
+app = Flask(__name__)
+load_config(app)
+app.register_blueprint(api, url_prefix='/qujata-api')
+
+CORS(app,origin=['*'])
+app.config['CORS_HEADERS'] = 'Content-Type'
+
+if __name__ == '__main__':
+    app.run(debug=True,host='0.0.0.0',port=3020)
