+++ portal/nginx.conf
@@ -14,12 +14,14 @@ http {
         server_tokens off;
 
         listen  80;
-        root    /usr/share/nginx/html/qujata;
+        root    /usr/share/nginx/html/;
         include /etc/nginx/mime.types;
 
         location /qujata {
-            try_files $uri $uri/ /index.html;
+            try_files $uri $uri/index.html =404;
         }
+        
+       
 
         # Cache for media: images, icons, fonts
         location ~* \.(?:ico|jpe?g|gif|png|ico|svg|woff2|woff)$ {
