+++ run/kubernetes/charts/api/values.yaml
@@ -49,13 +49,12 @@ ingress:
   enabled: true
   className: "nginx"
   annotations: 
-    nginx.ingress.kubernetes.io/rewrite-target: /$2
     "nginx.ingress.kubernetes.io/proxy-read-timeout": "3600"
     "nginx.ingress.kubernetes.io/proxy-send-timeout": "3600"
   hosts:
     - host: pqc.dev.att.com
       paths:
-        - path: /qujata-api(/|$)(.*)
+        - path: /qujata-api
           pathType: Prefix
   tls: 
     - secretName: tls-secret
