+++ run/kubernetes/charts/portal/values.yaml
@@ -37,7 +37,8 @@ securityContext: {}
   # runAsUser: 1000
 env:
   PQC_PORTAL__DASHBOARD_LINK_HOST: http://localhost:3000/grafana
-  PQC_PORTAL__BASE_API_URL: http://localhost:3020
+  PQC_PORTAL__BASE_API_URL: http://localhost:3020/qujata-api
+
 
 
 service:
